[Refactor] Move ti.parallel_sort under _kernels
